Add links command with Windows support
Changes:
- Rename link and clean commands to links init and links clean
- Add Windows support to links init and links clean (copy, then read-only)
- Make links.json platform-aware so entries target linux and/or windows
- Add links diff to report drift between repo files and their copies,
  with a hint for which side is newer
- Add links backport to copy file contents back into the repo,
  refusing to overwrite a newer repo file unless --force is given

+++ cli/lib/common/links.py
@@ -2,11 +2,14 @@
 
 import json
 import os
+import shutil
+import stat
 from enum import Enum
 from typing import List  # , Self
 
 from lib.common.dir import Dir
 from lib.common.log import Log
+from lib.common.os import OperatingSystem
 
 # from typing_extensions import Self
 
@@ -45,6 +48,12 @@ class Link:
         return self.link_type == LinkType.DIRECTORY
 
     def create(self) -> None:
+        if OperatingSystem.get().is_windows():
+            self._create_windows()
+        else:
+            self._create_linux()
+
+    def _create_linux(self) -> None:
         dst_dir = os.path.dirname(self.dst)
 
         if not os.path.exists(dst_dir):
@@ -68,7 +77,34 @@ class Link:
         Log.info(f"Creating symlink", {"source": self.src, "target": self.dst})
         os.symlink(self.src, self.dst)
 
+    def _create_windows(self) -> None:
+        dst_dir = os.path.dirname(self.dst)
+
+        if not os.path.exists(dst_dir):
+            Log.info(f"Creating parent directory for copy at path {self.dst}")
+            os.makedirs(dst_dir, exist_ok=True)
+
+        if not os.path.isdir(dst_dir):
+            raise Exception(
+                f"Parent directory path for copy {self.dst} exists but is not a directory"
+            )
+
+        if os.path.exists(self.dst):
+            Log.info(f"Deleting existing file at path {self.dst}")
+            os.chmod(self.dst, stat.S_IWRITE)
+            os.remove(self.dst)
+
+        Log.info(f"Copying file", {"source": self.src, "target": self.dst})
+        shutil.copyfile(self.src, self.dst)
+        os.chmod(self.dst, stat.S_IREAD)
+
     def delete(self) -> None:
+        if OperatingSystem.get().is_windows():
+            self._delete_windows()
+        else:
+            self._delete_linux()
+
+    def _delete_linux(self) -> None:
         if not os.path.islink(self.dst):
             if os.path.exists(self.dst):
                 Log.warn(f"File at path {self.dst} is not a symbolic link, skipping")
@@ -77,6 +113,15 @@ class Link:
             return
 
         Log.info(f"Removing symlink at path {self.dst}")
+        os.remove(self.dst)
+
+    def _delete_windows(self) -> None:
+        if not os.path.exists(self.dst):
+            Log.info(f"File at path {self.dst} does not exist, skipping")
+            return
+
+        Log.info(f"Removing file at path {self.dst}")
+        os.chmod(self.dst, stat.S_IWRITE)
         os.remove(self.dst)
 
 
@@ -95,6 +140,15 @@ class Links:
             return json.loads(f.read())
 
     @staticmethod
+    def _platform() -> str:
+        os_ = OperatingSystem.get()
+        if os_.is_windows():
+            return "windows"
+        if os_.is_linux():
+            return "linux"
+        return os_.get_name()
+
+    @staticmethod
     def _initialize_link(link_json) -> Link:
         return Link(
             os.path.join(Dir.config(), link_json["src"]),
@@ -104,6 +158,9 @@ class Links:
 
     @staticmethod
     def _initialize_links():
+        platform = Links._platform()
         Links._links = [
-            Links._initialize_link(link) for link in Links._load_links_json()
+            Links._initialize_link(link)
+            for link in Links._load_links_json()
+            if platform in link.get("platforms", ["linux"])
         ]
